requirements.txt
# Core Streamlit and UI
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.17.0
pandas>=2.1.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
langchain-experimental>=0.0.45

# Local LLM support
ollama>=0.1.0

# DevOps Tools Integration
docker>=6.1.0
kubernetes>=28.1.0
GitPython>=3.1.40

# Security Tools
bandit>=1.7.5
safety>=2.3.0
semgrep

# Code Analysis
pylint>=3.0.0
flake8>=6.1.0
mypy>=1.7.0
radon>=6.0.1

# Testing Tools
pytest>=7.4.0
pytest-cov>=4.1.0
locust>=2.17.0

# Infrastructure as Code
terraform-compliance>=1.3.44
checkov>=3.0.0

# Monitoring and Observability
prometheus-client>=0.19.0
grafana-api>=1.0.3

# Database
db-sqlite3

# Utilities
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
typer>=0.9.0

# Development
black>=23.11.0
isort>=5.12.0
pre-commit>=3.5.0
